Update the Refactoring project to Use React hooks
