+++ src/App.js
@@ -1,11 +1,10 @@
 import './App.css';
 import React from 'react';
-// import Calculator from './components/Calculator';
-import Calculator2 from './components/Calculator2';
+import Calculator from './components/Calculator';
 
 const App = () => (
   <div className="App">
-    <Calculator2 />
+    <Calculator />
   </div>
 );
